requirements.txt
aiogram==2.12.1
aiohttp==3.7.4.post0
aiohttp-healthcheck==1.3.1
aiojobs==0.3.0
aiomysql==0.0.21
aioredis==1.3.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
colorama==0.4.4
emoji==1.4.1
greenlet==1.1.0
hiredis==2.0.0
idna==3.2
importlib-metadata==4.6.1
loguru==0.5.3
multidict==4.7.6
parse==1.19.0
pyaes==1.6.1
pyasn1==0.4.8
PyMySQL==0.9.3
python-dateutil==2.8.2
pytz==2021.1
qiwipy==2.1.6
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.22
Telethon==1.23.0
typing-extensions==3.10.0.0
urllib3==1.26.6
win32-setctime==1.0.3
yarl==1.6.3
zipp==3.5.0
